+++ assets/integration/livecanvas/modules/plain-classes/main.js
@@ -1,0 +1,272 @@
+/**
+ * @module plain-classes 
+ * @package WindPress
+ * @since 3.0.0
+ * @author Joshua Gugun Siagian <suabahasa@gmail.com>
+ * 
+ * Enhance the plain classes on the properties panel.
+ */
+
+import './style.scss';
+
+import { logger } from '@/integration/common/logger';
+import { previewIframe } from '@/integration/livecanvas/constant.js';
+
+import tippy, { followCursor } from 'tippy.js';
+
+import { nextTick, ref, watch } from 'vue';
+import autosize from 'autosize';
+import Tribute from 'tributejs';
+
+import { createHighlighterCore } from 'shiki/core';
+import { createOnigurumaEngine } from 'shiki/engine/oniguruma';
+
+import HighlightInTextarea from '@/integration/library/highlight-in-textarea';
+
+import { debounce } from 'lodash-es';
+
+let shikiHighlighter = null;
+
+(async () => {
+    shikiHighlighter = await createHighlighterCore({
+        themes: [
+            import('shiki/themes/dark-plus.mjs'),
+            import('shiki/themes/light-plus.mjs'),
+        ],
+        langs: [
+            import('shiki/langs/css.mjs'),
+        ],
+        engine: createOnigurumaEngine(import('shiki/wasm')),
+    });
+})();
+
+const textInput = document.querySelector('div#sidepanel>section[item-type="edit-properties"]>form>div.common-form-fields>div.common-form-fields-content textarea[attribute-name="class"]');
+
+// highlight any text except spaces and new lines
+let hit = new HighlightInTextarea(textInput, {
+    highlight: [
+        {
+            highlight: /(?<=\s|^)(?:(?!\s).)+(?=\s|$)/g,
+            className: 'word',
+        },
+        {
+            highlight: /(?<=\s)\s/g,
+            className: 'multispace',
+            blank: true,
+        },
+    ],
+});
+
+hit.handleInput();
+autosize(textInput);
+
+let autocompleteItems = [];
+
+wp.hooks.addAction('windpresslivecanvas-autocomplete-items-refresh', 'windpresslivecanvas', () => {
+    // wp hook filters. {value, color?, fontWeight?, namespace?}[]
+    autocompleteItems = wp.hooks.applyFilters('windpresslivecanvas-autocomplete-items', [], textInput.value);
+});
+
+wp.hooks.doAction('windpresslivecanvas-autocomplete-items-refresh');
+
+const tribute = new Tribute({
+    containerClass: 'windpresslivecanvas-tribute-container',
+
+    autocompleteMode: true,
+
+    // Limits the number of items in the menu
+    menuItemLimit: 50,
+
+    noMatchTemplate: '',
+
+    values: async function (text, cb) {
+        const filters = await wp.hooks.applyFilters('windpresslivecanvas-autocomplete-items-query', text, autocompleteItems);
+        cb(filters);
+    },
+
+    lookup: 'value',
+
+    itemClass: 'class-item',
+
+    // template
+    menuItemTemplate: function (item) {
+        let customStyle = '';
+
+        if (item.original.color !== undefined) {
+            customStyle += `background-color: ${item.original.color};`;
+        }
+
+        if (item.original.fontWeight !== undefined) {
+            customStyle += `font-weight: ${item.original.fontWeight};`;
+        }
+
+        return `
+            <span class="class-name" data-tribute-class-name="${item.original.value}">${item.string}</span>
+            <span class="class-hint" style="${customStyle}"></span>
+        `;
+    },
+});
+
+tribute.setMenuContainer = function (el) {
+    this.menuContainer = el;
+};
+
+const tributeEventCallbackOrigFn = tribute.events.callbacks;
+
+tribute.events.callbacks = function () {
+    return {
+        ...tributeEventCallbackOrigFn.call(this),
+        up: (e, el) => {
+            // navigate up ul
+            if (this.tribute.isActive && this.tribute.current.filteredItems) {
+                e.preventDefault();
+                e.stopPropagation();
+                let count = this.tribute.current.filteredItems.length,
+                    selected = this.tribute.menuSelected;
+
+                if (count > selected && selected > 0) {
+                    this.tribute.menuSelected--;
+                    this.setActiveLi();
+                } else if (selected === 0) {
+                    this.tribute.menuSelected = count - 1;
+                    this.setActiveLi();
+                    this.tribute.menu.scrollTop = this.tribute.menu.scrollHeight;
+                }
+                // previewTributeEventCallbackUpDown();
+            }
+        },
+        down: (e, el) => {
+            // navigate down ul
+            if (this.tribute.isActive && this.tribute.current.filteredItems) {
+                e.preventDefault();
+                e.stopPropagation();
+                let count = this.tribute.current.filteredItems.length - 1,
+                    selected = this.tribute.menuSelected;
+
+                if (count > selected) {
+                    this.tribute.menuSelected++;
+                    this.setActiveLi();
+                } else if (count === selected) {
+                    this.tribute.menuSelected = 0;
+                    this.setActiveLi();
+                    this.tribute.menu.scrollTop = 0;
+                }
+                // previewTributeEventCallbackUpDown();
+            }
+        },
+    };
+};
+
+tribute.attach(textInput);
+
+const currentActiveSelector = ref('');
+
+function onTextInputChanges() {
+    nextTick(() => {
+        try {
+            hit.handleInput();
+        } catch (error) { }
+        autosize.update(textInput);
+        // tribute.setMenuContainer(document.querySelector('div.hit-container'));
+        tribute.hideMenu();
+    });
+};
+
+const editPropertiesObserver = new MutationObserver((mutations) => {
+    mutations.forEach((mutation) => {
+        if (mutation.type === 'attributes' && mutation.attributeName === 'selector') {
+            currentActiveSelector.value = mutation.target.getAttribute('selector');
+        }
+    });
+});
+
+editPropertiesObserver.observe(document.querySelector('div#sidepanel>section[item-type="edit-properties"]'), {
+    attributes: true,
+    attributeFilter: ['selector'],
+});
+
+// watch selector changes
+watch(currentActiveSelector, (newVal, oldVal) => {
+    if (newVal !== oldVal) {
+        onTextInputChanges();
+    }
+});
+
+// Disabled until the feature is stable
+textInput.addEventListener('highlights-updated', function (e) {
+    hoverPreviewProvider();
+});
+
+// create a tippy instance that will be used to show the hover preview, but not yet shown
+let tippyInstance = tippy(document.createElement('div'), {
+    plugins: [followCursor],
+    allowHTML: true,
+    arrow: false,
+    duration: [500, 0],
+    followCursor: true,
+    trigger: 'manual',
+});
+
+function hoverPreviewProvider() {
+    const hitContainerEl = document.querySelector('.hit-container');
+
+    if (hitContainerEl === null) {
+        return;
+    }
+
+    tippyInstance.reference = hitContainerEl;
+
+    async function showTippy(markWordElement) {
+        const classname = markWordElement.textContent;
+        const generatedCssCode = await previewIframe.contentWindow.windpress.module.classnameToCss.generate(classname);
+        if (generatedCssCode === null || generatedCssCode.trim() === '') {
+            return null;
+        };
+
+        tippyInstance.setContent(shikiHighlighter.codeToHtml(generatedCssCode, {
+            lang: 'css',
+            theme: 'dark-plus',
+        }));
+
+        tippyInstance.show();
+    }
+
+    const currentMarkWordElement = ref(null);
+
+    const debouncedMousemoveHandler = debounce(function (event) {
+        // if (!settingsState('module.plain-classes.hover-preview-classes', true).value) {
+        //     return;
+        // }
+
+        const x = event.clientX;
+        const y = event.clientY;
+
+        // get all elements that overlap the mouse
+        const elements = document.elementsFromPoint(x, y);
+
+        // find the first `mark` element
+        const firstMarkWordElement = elements.find((element) => {
+            return element.matches('mark[class="word"]');
+        });
+
+        currentMarkWordElement.value = firstMarkWordElement || null;
+    }, 10);
+
+    // when mouse are entering the `.hit-container` element, get the coordinates of the mouse and check if the mouse is hovering the `mark` element
+    hitContainerEl.addEventListener('mousemove', debouncedMousemoveHandler);
+
+    hitContainerEl.addEventListener('mouseleave', function (event) {
+        debouncedMousemoveHandler.cancel();
+        currentMarkWordElement.value = null;
+    });
+
+    watch(currentMarkWordElement, (newVal, oldVal) => {
+        if (newVal && newVal !== oldVal) {
+            showTippy(newVal);
+        } else {
+            tippyInstance.hide();
+        }
+    });
+}
+
+logger('Module loaded!', { module: 'plain-classes' });